+++ db/migrate/20140207134248_add_description_to_export_card_configurations.rb
@@ -1,0 +1,5 @@
+class AddDescriptionToExportCardConfigurations < ActiveRecord::Migration
+  def change
+    add_column :export_card_configurations, :description, :text
+  end
+end